refactor(apps/dex-swap-indexer): rename files to conform to jellyfish convention
